+++ src/app/features/events/events-routing.module.ts
@@ -8,6 +8,7 @@ import { EventResolver } from '@resolvers/event.resolver';
 
 import { EventsComponent } from '@features/events/events.component';
 import { EventsFavComponent } from '@features/events/events-fav/events-fav.component';
+import { EventsAuditComponent } from '@features/events/events-audit/events-audit.component';
 import { EventsOwnComponent } from '@features/events/events-own/events-own.component';
 import { EventViewComponent } from '@features/events/event-view/event-view.component';
 import { EventConfigComponent } from '@features/events/event-config/event-config.component';
@@ -26,6 +27,12 @@ const routes: Routes = [
   {
     path: 'new',
     component: EventCreationComponent,
+    canActivate: [ AngularFireAuthGuard ],
+    data: { authGuardPipe: redirectUnauthorizedToLogin },
+  },
+  {
+    path: 'audit',
+    component: EventsAuditComponent,
     canActivate: [ AngularFireAuthGuard ],
     data: { authGuardPipe: redirectUnauthorizedToLogin },
   },