+++ src/app/pages/calendar/calendar-header/calendar-header.component.ts
@@ -1,20 +1,23 @@
-import { Component, Input, Output, EventEmitter } from '@angular/core';
+import { Component, Input, OnInit, OnDestroy, Output, EventEmitter } from '@angular/core';
 import { MatDialog, MatDialogConfig } from '@angular/material/dialog';
+import { AngularFireAuth } from '@angular/fire/compat/auth';
 
+import { Subscription } from 'rxjs';
 import { CalendarView } from 'angular-calendar';
 
 import { EventsSearchDialogComponent } from '@features/events/events-search-dialog/events-search-dialog.component';
-import { Base, IBase } from '@models/base';
 import { CalendarEntitiesDialogComponent } from '@pages/calendar/calendar-entities-dialog/calendar-entities-dialog.component';
 import { CalendarModeDialogComponent } from '@pages/calendar/calendar-mode-dialog/calendar-mode-dialog.component';
-import { SwalMessage, UtilsService } from '@services/utils.service';
+import { Base, IBase } from '@models/base';
+import { IUser } from '@models/user';
+import { UserService } from '@services/users.service';
 
 @Component({
   selector: 'app-calendar-header',
   templateUrl: './calendar-header.component.html',
   styleUrls: ['./calendar-header.component.scss'],
 })
-export class CalendarHeaderComponent {
+export class CalendarHeaderComponent implements OnInit, OnDestroy {
 
   @Input() view: CalendarView;
   @Input() entityId: string;
@@ -23,27 +26,45 @@ export class CalendarHeaderComponent {
 
   @Output() viewChange = new EventEmitter<CalendarView>();
   @Output() viewDateChange = new EventEmitter<Date>();
-  @Output() entitySelectedChange = new EventEmitter<string[]>();
+  @Output() filterEntitiesChange = new EventEmitter<string[]>();
 
   public dialogConfig = new MatDialogConfig();
   CalendarView = CalendarView;
   modeSelected: string;
 
+  userLogged: IUser;
+  entities: string[] = [];
+  public favEntities: string[] = [];
   entityFilteredOption: string;
   entityFiltered: IBase;
 
+  private listOfObservers: Array<Subscription> = [];
+
   constructor(
     public dialog: MatDialog,
-    private utilsSrv: UtilsService,
+    public auth: AngularFireAuth,
+    private userSrv: UserService,
   ) {}
+
+  ngOnInit(): void {
+
+    const subs1$ = this.auth.user.subscribe((usr) => {
+      const uidUser = usr?.uid;
+      this.userSrv.getOneUser(uidUser)
+      .subscribe( (user: IUser) => {
+        this.userLogged = user;
+        this.favEntities = this.userLogged.favEntities;
+      });
+    });
+
+    this.listOfObservers.push(subs1$);
+  }
 
   gotoEventSearch() {
     const dialogRef1 = this.dialog.open(EventsSearchDialogComponent);
   }
 
   gotoModeConfig() {
-
-    console.log('gotoModeConfig');
 
     this.dialogConfig.data = { view: this.view };
     this.dialogConfig.width = '500px';
@@ -77,37 +98,37 @@ export class CalendarHeaderComponent {
 
     gotoEntitiesConfig() {
 
-      console.log(`gotoEntitiesConfig: ${this.entityFilteredOption} | ${this.entityFiltered?.name}`);
-
       this.dialogConfig.data = {
           option: this.entityFilteredOption,
-          entity: this.entityFiltered
+          entity: this.entityFiltered,
+          favEntities: this.favEntities,
       };
       this.dialogConfig.width = '500px';
       this.dialogConfig.height = '500px';
+
+      // console.log(`favEntities: ${this.favEntities.length}`);
 
       const dialogRef2 = this.dialog.open(CalendarEntitiesDialogComponent, this.dialogConfig);
 
       dialogRef2.afterClosed().subscribe(([entityOption, entitySelected]: [string, IBase]) => {
 
           this.entityFilteredOption = entityOption;
-          console.log(`closing - entityFilteredOption -> ${this.entityFilteredOption};`);
-
           this.entityFiltered = entitySelected?.id !== '0' ? entitySelected : null;
 
           if ( this.entityFilteredOption ) {
 
-              let entities = [];
+              this.entities = [];
               switch (this.entityFilteredOption) {
                 case '1': // cualquiera
-                  entities = [];
+                  this.entities = [];
                   break;
-                case '2': // TODO favoritos - TODO
+                case '2': // favoritos
+                  this.entities = this.favEntities;
                   break;
                 case '3': // una concreta
-                  entities = [entitySelected.id];
+                  this.entities = [entitySelected.id];
               }
-              this.entitySelectedChange.emit(entities);
+              this.filterEntitiesChange.emit(this.entities);
           } else {
             // this.utilsSrv.swalFire(SwalMessage.NO_CHANGES);
           }
@@ -119,5 +140,9 @@ export class CalendarHeaderComponent {
           background: `url('${this.entityFiltered.imagePath}') center no-repeat`,
         };
         return styles;
+    }
+
+    ngOnDestroy(): void {
+      this.listOfObservers.forEach(sub => sub.unsubscribe());
     }
 }