+++ src/environments/environment.ts
@@ -1,10 +1,10 @@
 export const environment = {
   production: false,
-  appVersion: '2.0.1.038',
-  releaseDate: '2022-05-28 02:00',
+  appVersion: '2.0.1.041',
+  releaseDate: '2022-06-03 00:00',
   baseUrl: 'https://agendarinconera.es',
   // baseUrl: 'http://localhost:4200',
-  setAudit: false,
+  setAudit: true,
   storiesNDaysAhead: 3,
   loginWithLogin: false,
   firebase: {